backend/pyproject.toml
[project]
name = "anonymous-chat-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
